Wip
Wip

fix

fix more tests

wip

Wips

partial

Fixes, lints

Lints
