+++ torch/_dynamo/config.py
@@ -103,6 +103,7 @@ if HAS_REFS_PRIMS:
         torch._refs,
         torch._prims,
         torch._decomp,
+        # torch._prims_common,
     }
 
 # If a string representing a PyTorch module is in this ignorelist,
