+++ torch/_dynamo/variables/base.py
@@ -1,10 +1,27 @@
 import collections
+import functools
+import numbers
+import operator
 from typing import Any, Callable, Dict, List, Optional, Set
+
+import torch
+from torch.fx.immutable_collections import immutable_list
 
 from .. import variables
 from ..exc import unimplemented
 from ..source import AttrSource, Source
-from ..utils import dict_values, identity, istype, odict_values
+from ..utils import (
+    clone_input,
+    dict_values,
+    fake_tensors_available,
+    get_fake_value,
+    get_real_value,
+    identity,
+    istype,
+    odict_values,
+    preserve_rng_state,
+    wrap_to_fake_tensor_and_record,
+)
 
 
 class MutableLocal:
@@ -19,6 +36,203 @@ class MutableLocal:
 
     def __eq__(self, other):
         return self is other
+
+
+def wrap_fx_proxy(tx, proxy, example_value=None, **options):
+    from .tensor import TensorVariable
+
+    return wrap_fx_proxy_cls(
+        target_cls=TensorVariable,
+        tx=tx,
+        proxy=proxy,
+        example_value=example_value,
+        **options,
+    )
+
+
+# Note: Unfortunate split due to some gross classes existing that subclass TensorVariable
+# Should be compositional instead
+def wrap_fx_proxy_cls(target_cls, tx, proxy, example_value=None, **options):
+    from .. import config, variables
+    from .constant import ConstantVariable
+    from .lists import SizeVariable
+    from .tensor import DynamicShapeVariable  # TODO(voz): Give it its own file
+
+    if "guards" in options and options["guards"] is not None:
+        tx.output.guards.update(options["guards"])
+
+    assert "example_value" not in proxy.node.meta
+    if not config.dynamic_propagation:
+        if isinstance(example_value, torch.Tensor):
+            options.update(target_cls.specialize(example_value))
+        return target_cls(proxy, **options)
+
+    use_fake_tensors = fake_tensors_available and config.fake_tensor_propagation
+
+    initial_example_value = example_value
+
+    def _clone_input(value):
+        if isinstance(value, torch.Tensor):
+            use_fake_tensors = fake_tensors_available and config.fake_tensor_propagation
+            # tensor subclasses will not be converted to FakeTensors and need to be cloned
+            if not use_fake_tensors or not isinstance(
+                value, torch._subclasses.fake_tensor.FakeTensor
+            ):
+                # NB: ensure strides are preserved
+                value = clone_input(value)
+
+        return value
+
+    with preserve_rng_state():
+        if example_value is None:
+            if use_fake_tensors:
+                example_value = get_fake_value(proxy.node, tx)
+            else:
+                example_value = get_real_value(proxy.node, tx.output)
+
+        else:
+            proxy.tracer.real_value_cache[proxy.node] = _clone_input(example_value)
+            if use_fake_tensors:
+                fake_wrapper = functools.partial(wrap_to_fake_tensor_and_record, tx=tx)
+                example_value = fake_wrapper(example_value)
+
+    if isinstance(example_value, torch.Tensor):
+        is_parameter = isinstance(example_value, torch.nn.Parameter)
+        should_specialize = options.pop("should_specialize", False)
+        if is_parameter or should_specialize:
+            specialized_value = initial_example_value
+        else:
+            specialized_value = None
+
+        example_value = _clone_input(example_value)
+        proxy.node.meta["example_value"] = example_value
+        specialized_props = target_cls.specialize(example_value)
+        if use_fake_tensors and isinstance(
+            example_value, torch._subclasses.fake_tensor.FakeTensor
+        ):
+            specialized_props["class_type"] = (
+                torch.nn.Parameter if is_parameter else torch.Tensor
+            )
+
+        specialized_props["specialized_value"] = specialized_value
+
+        options.update(specialized_props)
+        return target_cls(proxy, **options)
+    elif (
+        hasattr(proxy.node.target, "__name__")
+        and proxy.node.target.__name__ == "set_state"
+        and isinstance(proxy.node.target.__self__, torch._C.Generator)
+        or proxy.node.target == torch.random.set_rng_state
+    ):
+        from . import TorchVariable
+
+        return TorchVariable(proxy.node.target)
+    elif istype(example_value, (int, bool, float)) and not config.dynamic_shapes:
+        return ConstantVariable(example_value)
+    elif istype(example_value, (int, bool, float)) and config.dynamic_shapes:
+        proxy.node.meta["example_value"] = example_value
+        return DynamicShapeVariable.create(tx, proxy, example_value, **options)
+    elif istype(example_value, torch.Size) and config.dynamic_shapes:
+        proxy.node.meta["example_value"] = example_value
+        sizes = []
+        for i, v in enumerate(example_value):
+            proxy_i = proxy[i]
+            sizes.append(DynamicShapeVariable.create(tx, proxy_i, v, **options))
+        return SizeVariable(sizes, proxy, **options)
+    elif istype(example_value, int) and proxy.node.target in (
+        torch.seed,
+        operator.mod,
+        # some mac builds are missing torch.distributed.get_rank()
+        getattr(torch.distributed, "get_rank", _missing),
+        getattr(torch.distributed, "get_world_size", _missing),
+    ):
+        proxy.node.meta["example_value"] = example_value
+        return DynamicShapeVariable.create(tx, proxy, example_value, **options)
+    elif istype(example_value, torch.Size) and all(
+        [isinstance(x, int) for x in example_value]
+    ):
+        sizes = [variables.ConstantVariable(x) for x in example_value]
+        return SizeVariable(sizes, **options)
+    elif isinstance(example_value, (tuple, list)):
+        unpacked = []
+        for i, val in enumerate(example_value):
+            if val is None:
+                # nn.MultiheadAttention() can return None, see issue #175
+                unpacked.append(
+                    variables.ConstantVariable(None, **options),
+                )
+            else:
+                unpacked.append(
+                    wrap_fx_proxy(
+                        tx,
+                        proxy.tracer.create_proxy(
+                            "call_function", operator.getitem, (proxy, i), {}
+                        ),
+                        example_value=val,
+                        **options,
+                    )
+                )
+        if istype(example_value, tuple):
+            return variables.TupleVariable(unpacked, **options)
+        elif istype(example_value, (list, immutable_list)):
+            return variables.ListVariable(
+                unpacked, mutable_local=MutableLocal(), **options
+            )
+        else:
+            assert (
+                example_value.__class__.__module__ == "torch.return_types"
+                or hasattr(example_value, "_fields")
+            ), ("namedtuple?")
+            return variables.NamedTupleVariable(
+                unpacked, example_value.__class__, **options
+            )
+    elif example_value is None or proxy.node.target is torch.manual_seed:
+        return variables.ConstantVariable(None, **options)
+    elif (
+        isinstance(example_value, int)
+        and proxy.node.target is torch._utils._element_size
+    ):
+        proxy.node.meta["example_value"] = example_value
+        return variables.ConstantVariable(example_value, **options)
+    elif (
+        isinstance(example_value, numbers.Number)
+        and (proxy.node.target == "item" or proxy.node.target in {math.sqrt, math.pow})
+        and config.capture_scalar_outputs
+    ):
+        if use_fake_tensors:
+            # item raw value should not be accessed
+            return wrap_fx_proxy_cls(
+                FakeItemVariable,
+                tx=tx,
+                proxy=proxy,
+                example_value=torch.tensor(example_value),
+                **options,
+            )
+        else:
+            return wrap_fx_proxy_cls(
+                UnspecializedPythonVariable,
+                tx=tx,
+                proxy=proxy,
+                example_value=torch.tensor(example_value),
+                raw_value=None if use_fake_tensors else example_value,
+                need_unwrap=False,
+                **options,
+            )
+    elif (
+        proxy.node.target == torch._C._DisableFuncTorch
+        or proxy.node.target == torch.cuda._is_in_bad_fork
+    ):
+        from . import UserDefinedObjectVariable
+
+        return UserDefinedObjectVariable(example_value)
+    elif isinstance(example_value, (torch.SymInt, torch.SymFloat)):
+        proxy.node.meta["example_value"] = example_value
+        return DynamicShapeVariable(proxy, example_value, **options)
+    else:
+        raise AssertionError(
+            "torch.* op returned non-Tensor "
+            + f"{typestr(example_value)} {proxy.node.op} {proxy.node.target}"
+        )
 
 
 class VariableTracker:
