+++ torch/_dynamo/variables/builder.py
@@ -42,7 +42,7 @@ from ..utils import (
     tuple_iterator_getitem,
     tuple_iterator_len,
 )
-from .base import MutableLocal
+from .base import MutableLocal, wrap_fx_proxy, wrap_fx_proxy_cls
 from .builtin import BuiltinVariable
 from .constant import ConstantVariable, EnumVariable
 from .dicts import (
@@ -72,7 +72,6 @@ from .misc import (
 )
 from .nn_module import UnspecializedNNModuleVariable
 from .tensor import (
-    TensorVariable,
     TensorWithTFOverrideVariable,
     UnspecializedNumpyVariable,
     UnspecializedPythonVariable,
@@ -536,7 +535,7 @@ class VariableBuilder:
                         source=None,
                         # Guards are added inside register_attr_or_module
                     )
-                tensor_variable = TensorVariable.create(
+                tensor_variable = wrap_fx_proxy(
                     tx=self.tx,
                     proxy=self.tx.output.create_graph_input(
                         re.sub(r"[^a-zA-Z0-9]+", "_", self.name), type(value)
@@ -578,14 +577,16 @@ class VariableBuilder:
             )
 
             if isinstance(value, np.number):
-                unspec_var = UnspecializedNumpyVariable.create(
+                unspec_var = wrap_fx_proxy_cls(
+                    UnspecializedNumpyVariable,
                     tx=self.tx,
                     proxy=proxy,
                     example_value=wrapped_value,
                     **options,
                 )
             else:
-                unspec_var = UnspecializedPythonVariable.create(
+                unspec_var = wrap_fx_proxy_cls(
+                    UnspecializedPythonVariable,
                     tx=self.tx,
                     proxy=proxy,
                     example_value=wrapped_value,
